fix: change double saldo for int, as the prompt

+++ modulo8-POO/M8E3-Banco/src/Cuenta.java
@@ -1,9 +1,9 @@
 public class Cuenta {
     // Atributos
     private int nroCuenta;
-    private double saldo;
+    private int saldo;
     // Constructor
-    public Cuenta (int nroCuenta, double saldo) {
+    public Cuenta (int nroCuenta, int saldo) {
         this.nroCuenta = nroCuenta;
         this.saldo = saldo;
     }
@@ -12,23 +12,23 @@ public class Cuenta {
     public int getNroCuenta() {
         return nroCuenta;
     }
-    public double getSaldo() {
+    public int getSaldo() {
         return saldo;
     }
     // Setters
     public void setNroCuenta (int nroCuenta) {
         this.nroCuenta = nroCuenta;
     }
-    public void setSaldo (double saldo) {
+    public void setSaldo (int saldo) {
         this.saldo = saldo;
     }
     // Métodos particulares
     // Método para ingresar/depositar saldo
-    public double ingresar(double deposito) {
+    public int ingresar(int deposito) {
         return this.saldo = saldo + deposito;
     }
     // Método para retirar/extraer saldo
-    public double retirar(double extraccion) {
+    public int retirar(int extraccion) {
         return this.saldo = saldo - extraccion;
     }
